Moved tasks to better folders
