+++ packages/sequencer/src/worker/startup/CloseWorkerError.ts
