+++ packages/sequencer/src/worker/startup/WorkerRegistrationFlow.ts
@@ -1,8 +1,8 @@
 import { injectable } from "tsyringe";
 import { log } from "@proto-kit/common";
 
-import { Closeable } from "../../../sequencer/builder/Closeable";
-import { FlowCreator } from "../../flow/Flow";
+import { Closeable } from "../../sequencer/builder/Closeable";
+import { FlowCreator } from "../flow/Flow";
 
 import {
   WorkerRegistrationTask,
